libs/python/clients/apis/config-vault/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "cli_config_vault" ]

[tool.coverage.report]
exclude_lines = [ "if TYPE_CHECKING:" ]
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-fail-under=80 --cov-report html:'../../../../../coverage/libs/python/clients/apis/config-vault/html' --cov-report xml:'../../../../../coverage/libs/python/clients/apis/config-vault/coverage.xml' --html='../../../../../reports/libs/python/clients/apis/config-vault/unittests/html/index.html' --junitxml='../../../../../reports/libs/python/clients/apis/config-vault/unittests/junit.xml'"

[tool.poetry]
name = "python-clients-apis-config-vault"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = "Proprietary"
readme = "README.md"

  [[tool.poetry.packages]]
  include = "cli_config_vault"

  [tool.poetry.dependencies]
  python = "^3.10"

    [tool.poetry.dependencies.shared-py-request]
    path = "../../../shared/py-request"
    develop = true

    [tool.poetry.dependencies.python-service-discovery]
    path = "../../../service-discovery"
    develop = true

    [tool.poetry.dependencies.shared-py-serializer]
    path = "../../../shared/py-serializer"
    develop = true

    [tool.poetry.dependencies.python-ddd-dtos-config-vault]
    path = "../../../ddd/dtos/config-vault"
    develop = true

    [tool.poetry.dependencies.shared-py-log]
    path = "../../../shared/py-log"
    develop = true


[tool.poetry.group.dev.dependencies]
respx = "^0.21.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
